Day 14: High and Lower Game
Day 14/100
This repository is dedicated to the 100 Days of Code Challenge for 2023, with a focus on the Complete Python Pro Bootcamp. The challenge aims to provide an introduction to Python programming, including the basics of syntax, data structures, and control flow. The goal is to build a solid foundation in the language and provide the necessary tools for future challenges and projects.